refactor: 重构架构以支持多实例并行 SITL
- 新增 `icsearcher/concurrency.py` 用于多实例编排
- 在配置中增加 `parallel.instances` 选项及相关路径/端口分配
- 更新 `collect` 和 `validate` 流水线以支持并发执行
- 补充 `concurrency` 与 `io` 相关的测试用例
- 更新 README 文档与配置示例

+++ icsearcher/params.py
@@ -76,7 +76,10 @@ def read_range_from_dict(para_dict):
 
 
 def read_unit_from_dict(para_dict):
-    return para_dict.loc['step'].to_numpy()
+    # Coerce to float: the param JSON mixes integer and decimal step values,
+    # so pandas infers an object dtype, and ``param * step_unit`` then propagates
+    # object dtype — which breaks downstream (np.isfinite, model input, etc.).
+    return para_dict.loc['step'].to_numpy(dtype=float)
 
 
 # Log analysis function
